requirements.txt
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
psycopg2==2.7.5
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pyuploadcare==2.7.0
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
Pillow==5.2.0
importlib-metadata==1.6.0